+++ changelog.md
@@ -4,6 +4,7 @@
 - more editor functions as always
 - editor scale compat functions
 - update editor tab buttons in registeredittabbutton so it can b used post hook
+- include ranges header with utils
 
 ## v1.5.1
 **Fixed**